requirements Spaces.txt
torch
torchvision
gradio
pillow
numpy
scikit-learn
matplotlib
seaborn
tqdm
python-docx
datasets
huggingface_hub